Implementado método de login y obtención de cuentas
Se refactorizó y se añadieron métodos para la validación del inicio de sesión del usuario, la obtención de cuentas y la verificación de la cuenta de destino. Se añadieron nuevos modelos para el historial y los resultados de las transacciones, y se actualizaron las cuentas y las transacciones para usar decimales en los valores monetarios. Los archivos del proyecto y de recursos se actualizaron para reflejar estos cambios.

+++ Parcial2DesIV/Modelos/Transacciones.cs
@@ -11,7 +11,7 @@ namespace Parcial2DesIV.Modelos
         public int id { get; set; }
         public string id_cuenta_origen { get; set; }
         public string id_cuenta_destino { get; set; }
-        public float monto { get; set; }
+        public decimal monto { get; set; }
         public DateTime fecha { get; set; }
     }
 }
